+++ ocean_lib/assets/test/test_asset_downloader.py
@@ -3,15 +3,17 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 import os
+from unittest.mock import patch
 
 import pytest
 from requests.exceptions import InvalidURL
 
-from ocean_lib.agreements.consumable import AssetNotConsumable
+from ocean_lib.agreements.consumable import AssetNotConsumable, ConsumableCodes
 from ocean_lib.agreements.service_types import ServiceTypes
 from ocean_lib.assets.asset import Asset
-from ocean_lib.assets.asset_downloader import download_asset_files
+from ocean_lib.assets.asset_downloader import download_asset_files, is_consumable
 from ocean_lib.data_provider.data_service_provider import DataServiceProvider
+from ocean_lib.services.service import Service
 from ocean_lib.web3_internal.currency import to_wei
 from tests.resources.ddo_helpers import (
     create_asset,
@@ -19,6 +21,30 @@ from tests.resources.ddo_helpers import (
     get_first_service_by_type,
     get_sample_ddo,
 )
+
+
+@pytest.mark.unit
+def test_is_consumable():
+    ddo_dict = get_sample_ddo()
+    asset = Asset.from_dict(ddo_dict)
+    service_dict = ddo_dict["services"][0]
+    service = Service.from_dict(service_dict)
+    with patch(
+        "ocean_lib.assets.test.test_asset_downloader.DataServiceProvider.check_asset_file_info",
+        return_value=False,
+    ):
+        assert (
+            is_consumable(asset, service, {}, True) == ConsumableCodes.CONNECTIVITY_FAIL
+        )
+
+    with patch(
+        "ocean_lib.assets.test.test_asset_downloader.DataServiceProvider.check_asset_file_info",
+        return_value=True,
+    ):
+        assert (
+            is_consumable(asset, service, {"type": "address", "value": "0xdddd"}, True)
+            == ConsumableCodes.CREDENTIAL_NOT_IN_ALLOW_LIST
+        )
 
 
 @pytest.mark.unit
